# specs/002-fullstack-web-app/contracts/api.yaml
openapi: 3.1.0
info:
  title: Todo App API
  description: RESTful API for the Full-Stack Web Todo App. All endpoints require JWT Bearer authentication.
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Local development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Better Auth on the frontend

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (required)
        description:
          type: string
          maxLength: 1000
          default: ""
          description: Task description (optional)

    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Updated task title (required)
        description:
          type: string
          maxLength: 1000
          default: ""
          description: Updated task description (optional)

    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Task description
        completed:
          type: boolean
          description: Whether the task is completed
        user_id:
          type: string
          description: Owner user ID (from JWT)
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (UTC)

    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskResponse"

    MessageResponse:
      type: object
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message

security:
  - BearerAuth: []

paths:
  /api/{user_id}/tasks:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT subject)

    get:
      summary: List all tasks for user
      operationId: listTasks
      tags: [Tasks]
      description: Returns all tasks belonging to the authenticated user. User ID in path must match the JWT token subject.
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User ID in path does not match JWT subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Create a new task
      operationId: createTask
      tags: [Tasks]
      description: Creates a new task for the authenticated user. Title is required (1-200 chars). Description is optional (max 1000 chars).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User ID in path does not match JWT subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error (title required, length limits)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/{user_id}/tasks/{task_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT subject)
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Task UUID

    get:
      summary: Get task details
      operationId: getTask
      tags: [Tasks]
      description: Returns details of a specific task. Task must belong to the authenticated user.
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User ID mismatch or task belongs to another user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update a task
      operationId: updateTask
      tags: [Tasks]
      description: Updates the title and/or description of a task. Does not change completion status (use PATCH /complete for that).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User ID mismatch or task belongs to another user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete a task
      operationId: deleteTask
      tags: [Tasks]
      description: Permanently removes a task. Task must belong to the authenticated user.
      responses:
        "200":
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User ID mismatch or task belongs to another user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/{user_id}/tasks/{task_id}/complete:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT subject)
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Task UUID

    patch:
      summary: Toggle task completion
      operationId: toggleTaskComplete
      tags: [Tasks]
      description: Toggles the completion status of a task. If incomplete, marks as complete. If complete, marks as incomplete.
      responses:
        "200":
          description: Task completion toggled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User ID mismatch or task belongs to another user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"